server/requirements.txt
flask
flask-cors
openai
python-dotenv
crewai
langchain
langchain-community
